Follow black tape with reflectance sensors
Added the code to use the reflectance sensors on their own to follow a piece of black tape on the table. I added this as a new Auto mode called, “Follow Black Auto w/ Sensors”.

I also learned how to add variables to the SmartDashboard to monitor variables as the robot moves along.